rename: made navigation plural
